War wagon, spearmen, and textures
Added Generic War Wagon. It uses the texture of the infantry for the crew so I had to modify it to include all the wagon textures. So I modified all the infantry URLs textures.

I also added spearmen with shields. The un-shielded ones were made to represent irish kern (although I am using it to all the 3Sp bases for now), and the new spearmen shielded are more like european militia. I already had the models, just uploaded them and added to this file.

And last, I added a new campament for the late german army I made. It has the sacred empire banner (although the blue player still uses the generic one).