Several mayor changes
All the 14 army lists included, with al least a bannerman for each army.

Bannerman's banner corrected. There were several polygons missplaced. Now it is correct and, since each side of the banner uses the same patron, I moved the UVs from both faces to the same area. This means we have extra space. I was thinking about including an extra banner in the future (different from the first we already have) with extra bannerman models with the banner pointing that second flag, just for variety purposes.

Included 2 extra models for the infantry bannerman and 1 extra for the knight bannerman. 

I reordered the armies, and now they are in the same order the army lists book IV has. I left generic late medieval on the top of the file.

Included a single texture for each army's bannerman, both infantry and knight. The only change in the texture is the banner so far. Livery changes will come later.

Included all the banners' references and rights, most of them are CC 3.0 or Public domain.